requirements.txt
Flask==2.3.2
Bootstrap-Flask==2.2.0
pillow == 10.2.0
Flask-WTF==1.2.1
scikit-learn == 1.4.0
numpy == 1.26.3
matplotlib == 3.8.2